変更

+++ app/Http/Controllers/TaskController.php
@@ -76,7 +76,9 @@ class TaskController extends Controller
        */
    public function destroy(Request $request, Task $task)
    {
-       $task->delete();
-       return redirect('/tasks');
+        $this->authorize('destroy',$task);
+
+        $task->delete();
+        return redirect('/tasks');
    }
 }
